requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-huggingface
langchain-experimental
langchain-tavily
python-dotenv
pydantic
duckduckgo-search
pypdf
tiktoken
langgraph
grandalf
pandas
numpy
scikit-learn
tabulate
missingpy
tensorly
pyppeteer